backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
asyncpg
python-jose[cryptography]
passlib[argon2]
python-multipart
httpx
pydantic-settings
argostranslate
redis
